Update license file with GPL3 text
The license in the COPYING file is GNU Lesser General Public License 3
while the headers of each file of the source indicate that they're
licensed under GPL 3. I emailed the original author on January 14, 2023
and asked for clarification on this discrepancy. He responded the same
day stating that the software is licensed under GPL 3.